工具条增加编辑与删除按钮

+++ admin-web/src/main/webapp/admin/user/controller/UserGridController.js
@@ -36,36 +36,48 @@ Ext.define('Kalix.admin.user.controller.UserGridController', {
      * @param rowIndex
      * @param colIndex
      */
-    onEdit: function (grid, rowIndex, colIndex) {
-        var rec = grid.getStore().getAt(rowIndex);
-        var editFormPanel = Ext.create('Kalix.admin.user.view.UserEditForm', {
-            url: this.getView().getViewModel().get("url")
-        });
-        editFormPanel.down("#confirmPasswordId").setValue(rec.data.password);
-        editFormPanel.loadRecord(rec);
+    onEdit: function () {
+        var grid = Ext.ComponentQuery.query('userGridPanel')[0];
+        var selModel = grid.getSelectionModel();
+        if (selModel.hasSelection()) {
+            var rows = selModel.getSelection();
+            if (rows.length > 1) {
+                Ext.Msg.alert(CONFIG.ALTER_TITLE_ERROR, "请选择一条要编辑的记录！");
+                return;
+            }
+            var rec = rows[0];
 
-        var win = Ext.create('Ext.Window', {
-            width: 400,
-            height: 350,
-            border: false,
-            modal: true,
-            //resizable:false,
-            icon: 'admin/resources/images/group_edit.png',
-            title: this.getView().getViewModel().get("editTitle"),
-            items: [editFormPanel]
-        });
+            var editFormPanel = Ext.create('Kalix.admin.user.view.UserEditForm', {
+                url: this.getView().getViewModel().get("url")
+            });
+            editFormPanel.down("#confirmPasswordId").setValue(rec.data.password);
+            editFormPanel.loadRecord(rec);
 
-        win.show();
+            var win = Ext.create('Ext.Window', {
+                width: 400,
+                height: 350,
+                border: false,
+                modal: true,
+                //resizable:false,
+                icon: 'admin/resources/images/group_edit.png',
+                title: this.getView().getViewModel().get("editTitle"),
+                items: [editFormPanel]
+            });
+
+            win.show();
+        } else {
+            Ext.Msg.alert(CONFIG.ALTER_TITLE_ERROR, "请选择要编辑的记录！");
+        }
     },
     /**
      * 批量删除操作.
      */
     onDeleteAll: function () {
-
         var selModel = Ext.ComponentQuery.query('#userDataGrid').getSelectionModel();
         if (selModel.hasSelection()) {
             Ext.Msg.confirm("警告", "确定要删除吗？", function (button) {
                 if (button == "yes") {
+                    var deleteUrl = this.getView().getViewModel().get("url");
                     var rows = selModel.getSelection();
                     var ids = "";
                     for (var i = 0; i < rows.length; i++) {
@@ -77,28 +89,14 @@ Ext.define('Kalix.admin.user.controller.UserGridController', {
                         }
                     }
                     Ext.Ajax.request({
-                        url: "/userDeleteAllServlet?ids=" + ids,
-                        method: "GET",
+                        url: deleteUrl + "?id=" + rec.id,
+                        method: 'DELETE',
                         callback: function (options, success, response) {
                             var resp = Ext.JSON.decode(response.responseText);
                             Ext.MessageBox.alert(CONFIG.ALTER_TITLE_INFO, resp.msg);
                             if (resp.success) {
-                                //var username = Ext.getCmp("username").getValue();
-                                //var name = Ext.getCmp("name").getValue();
-                                //var sex = Ext.getCmp("sex").getValue();
-                                //var status = Ext.getCmp("status").getValue();
-                                //var grid = Ext.getCmp("userDataGrid");
-                                //var store = grid.getStore();
-                                //store.reload({
-                                //    params: {
-                                //        start: 0,
-                                //        limit: pageSize,
-                                //        username: username,
-                                //        name: name,
-                                //        sex: sex,
-                                //        status: status
-                                //    }
-                                //});
+                                var store = grid.getStore();
+                                store.reload();
                             }
                         }
                     });
@@ -114,24 +112,35 @@ Ext.define('Kalix.admin.user.controller.UserGridController', {
      * @param rowIndex
      * @param colIndex
      */
-    onDelete: function (grid, rowIndex, colIndex) {
-        var rec = grid.getStore().getAt(rowIndex);
+    onDelete: function () {
+        var grid = Ext.ComponentQuery.query('userGridPanel')[0];
+        var selModel = grid.getSelectionModel();
         var deleteUrl = this.getView().getViewModel().get("url");
-        Ext.Msg.confirm("警告", "确定要删除吗？", function (button) {
-            if (button == "yes") {
-                Ext.Ajax.request({
-                    url: deleteUrl + "?id=" + rec.id,
-                    method: 'DELETE',
-                    callback: function (options, success, response) {
-                        var resp = Ext.JSON.decode(response.responseText);
-                        Ext.MessageBox.alert(CONFIG.ALTER_TITLE_INFO, resp.msg);
-                        if (resp.success) {
-                            var store = grid.getStore();
-                            store.reload();
-                        }
-                    }
-                });
+        if (selModel.hasSelection()) {
+            var rows = selModel.getSelection();
+            if (rows.length > 1) {
+                Ext.Msg.alert(CONFIG.ALTER_TITLE_ERROR, "请选择一条要删除的记录！");
+                return;
             }
-        });
+            var rec = rows[0];
+            Ext.Msg.confirm("警告", "确定要删除吗？", function (button) {
+                if (button == "yes") {
+                    Ext.Ajax.request({
+                        url: deleteUrl + "?id=" + rec.id,
+                        method: 'DELETE',
+                        callback: function (options, success, response) {
+                            var resp = Ext.JSON.decode(response.responseText);
+                            Ext.MessageBox.alert(CONFIG.ALTER_TITLE_INFO, resp.msg);
+                            if (resp.success) {
+                                var store = grid.getStore();
+                                store.reload();
+                            }
+                        }
+                    });
+                }
+            });
+        } else {
+            Ext.Msg.alert(CONFIG.ALTER_TITLE_ERROR, "请选择要删除的记录！");
+        }
     }
 });
